Some of the api fixes.
