+++ packages/office-ui-fabric-react/src/DragDrop.ts
@@ -1,0 +1,1 @@
+export * from 'office-ui-fabric-react/lib/DragDrop';
