+++ packages/office-ui-fabric-react/src/Keytips.ts
@@ -1,0 +1,4 @@
+export * from './components/Keytip/index';
+export * from './components/KeytipData/index';
+export * from './components/KeytipLayer/index';
+export * from './utilities/keytips/index';
