+++ packages/office-ui-fabric-react/src/Positioning.ts
@@ -1,0 +1,1 @@
+export * from './utilities/positioning/index';
